[ARM PATCH] 2557/1: S3C2410 - fix otom/nexcoder buiilds due to sparse fixes
Patch from Ben Dooks

These two machines where missed out when updating
for sparse fixes, return them to building correctly.

+++ arch/arm/mach-s3c2410/mach-otom.c
@@ -38,7 +38,7 @@
 
 static struct map_desc otom11_iodesc[] __initdata = {
   /* Device area */
-	{ OTOM_VA_CS8900A_BASE,  OTOM_PA_CS8900A_BASE,	   SZ_16M, MT_DEVICE },
+	{ (u32)OTOM_VA_CS8900A_BASE, OTOM_PA_CS8900A_BASE, SZ_16M, MT_DEVICE },
 };
 
 #define UCON S3C2410_UCON_DEFAULT
@@ -116,7 +116,7 @@ void __init otom11_map_io(void)
 
 MACHINE_START(OTOM, "Nex Vision - Otom 1.1")
      MAINTAINER("Guillaume GOURAT <guillaume.gourat@nexvision.tv>")
-     BOOT_MEM(S3C2410_SDRAM_PA, S3C2410_PA_UART, S3C24XX_VA_UART)
+     BOOT_MEM(S3C2410_SDRAM_PA, S3C2410_PA_UART, (u32)S3C24XX_VA_UART)
      BOOT_PARAMS(S3C2410_SDRAM_PA + 0x100)
 	.map_io		= otom11_map_io,
 	.init_irq	= s3c24xx_init_irq,
